# frontend/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Security AI Agent",
    "version": "0.1.0"
  },
  "paths": {
    "/api/v1/events": {
      "post": {
        "summary": "Submit Event",
        "operationId": "submit_event_api_v1_events_post",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "sync",
            "in": "query",
            "required": false,
            "schema": {
              "type": "boolean",
              "description": "sync mode",
              "default": false,
              "title": "Sync"
            },
            "description": "sync mode"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/EventSubmitRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EventSubmitResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "get": {
        "tags": [
          "operations"
        ],
        "summary": "List Events",
        "operationId": "list_events_api_v1_events_get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "status",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Status"
            }
          },
          {
            "name": "verdict",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Verdict"
            }
          },
          {
            "name": "priority",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Priority"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "default": 50,
              "title": "Limit"
            }
          },
          {
            "name": "offset",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "default": 0,
              "title": "Offset"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EventListResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "ApprovalEntry": {
        "properties": {
          "event_id": {
            "type": "string",
            "title": "Event Id"
          },
          "action": {
            "type": "string",
            "title": "Action"
          },
          "note": {
            "type": "string",
            "title": "Note",
            "default": ""
          },
          "actor": {
            "type": "string",
            "title": "Actor"
          },
          "role": {
            "type": "string",
            "title": "Role"
          },
          "timestamp": {
            "type": "string",
            "title": "Timestamp"
          }
        },
        "type": "object",
        "required": [
          "event_id",
          "action",
          "actor",
          "role",
          "timestamp"
        ],
        "title": "ApprovalEntry"
      },
...
      "EventListResponse": {
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/EventRecord"
            },
            "type": "array",
            "title": "Items"
          },
          "total": {
            "type": "integer",
            "title": "Total"
          }
        },
        "type": "object",
        "required": [
          "items",
          "total"
        ],
        "title": "EventListResponse"
      },
      "EventRecord": {
        "properties": {
          "event_id": {
            "type": "string",
            "title": "Event Id"
          },
          "source": {
            "type": "string",
            "title": "Source",
            "default": "api"
          },
          "submitted_at": {
            "type": "string",
            "title": "Submitted At",
            "default": ""
          },
          "finished_at": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Finished At"
          },
          "status": {
            "type": "string",
            "enum": [
              "processing",
              "completed",
              "pending_approval",
              "ignored",
              "error",
              "rejected"
            ],
            "title": "Status",
            "default": "processing"
          },
          "priority": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Priority"
          },
          "tags": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Tags",
            "default": []
          },
          "sanitized_text": {
            "type": "string",
            "title": "Sanitized Text",
            "default": ""
          },
          "iocs": {
            "additionalProperties": {
              "items": {
                "type": "string"
              },
              "type": "array"
            },
            "type": "object",
            "title": "Iocs",
            "default": {}
          },
          "final_verdict": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Final Verdict"
          },
          "confidence": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Confidence"
          },
          "evidence_summary": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Evidence Summary"
          },
          "mitre_ttps": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Mitre Ttps",
            "default": []
          },
          "poc": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Poc"
          },
          "is_vulnerable": {
            "type": "boolean",
            "title": "Is Vulnerable",
            "default": false
          },
          "trace": {
            "items": {
              "$ref": "#/components/schemas/TraceStep"
            },
            "type": "array",
            "title": "Trace",
            "default": []
          },
          "approvals": {
            "items": {
              "$ref": "#/components/schemas/ApprovalEntry"
            },
            "type": "array",
            "title": "Approvals",
            "default": []
          },
          "pending_approval_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Pending Approval Id"
          },
          "duration_ms": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Duration Ms"
          },
          "execution_summary": {
            "anyOf": [
              {
                "type": "object"
              },
              {
                "items": {},
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Execution Summary"
          }
        },
        "type": "object",
        "required": [
          "event_id"
        ],
        "title": "EventRecord"
      },
      "EventSubmitRequest": {
        "properties": {
          "sanitized_text": {
            "type": "string",
            "title": "Sanitized Text"
          },
          "iocs": {
            "additionalProperties": {
              "items": {
                "type": "string"
              },
              "type": "array"
            },
            "type": "object",
            "title": "Iocs"
          },
          "source": {
            "type": "string",
            "title": "Source",
            "default": "api"
          }
        },
        "type": "object",
        "required": [
          "sanitized_text",
          "iocs"
        ],
        "title": "EventSubmitRequest"
      },
      "EventSubmitResponse": {
        "properties": {
          "event_id": {
            "type": "string",
            "title": "Event Id"
          },
          "status": {
            "type": "string",
            "title": "Status"
          }
        },
        "type": "object",
        "required": [
          "event_id",
          "status"
        ],
        "title": "EventSubmitResponse"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "TraceStep": {
        "properties": {
          "node": {
            "type": "string",
            "title": "Node"
          },
          "action": {
            "type": "string",
            "title": "Action"
          },
          "summary": {
            "type": "string",
            "title": "Summary"
          },
          "timestamp": {
            "type": "string",
            "title": "Timestamp"
          },
          "details": {
            "type": "object",
            "title": "Details",
            "default": {}
          }
        },
        "type": "object",
        "required": [
          "node",
          "action",
          "summary",
          "timestamp"
        ],
        "title": "TraceStep"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
    },
    "securitySchemes": {
      "HTTPBearer": {
        "type": "http",
        "scheme": "bearer"
      }
    }
  }
}